FILE: moves filters component into filterable-list folder
