+++ libs/menu-matriarch/shared/ui-domain/src/lib/filterable-list/filterable-list.component.ts
@@ -18,7 +18,7 @@ import {
 } from '@atocha/menu-matriarch/shared/ui-generic';
 import { Tag } from '@atocha/menu-matriarch/shared/util';
 import { FiltersButtonComponent } from '../filters-button/filters-button.component';
-import { FiltersComponent } from '../filters/filters.component';
+import { FiltersComponent } from './filters/filters.component';
 
 @Component({
   standalone: true,
