+++ libs/menu-matriarch/shared/ui-domain/src/lib/filterable-list/filters/filters.component.ts
@@ -11,9 +11,9 @@ import { FormsModule } from '@angular/forms';
 import { ButtonComponent, CheckboxComponent } from '@atocha/core/ui';
 import { CardComponent } from '@atocha/menu-matriarch/shared/ui-generic';
 import { Tag } from '@atocha/menu-matriarch/shared/util';
-import { TagComponent } from '../tags/tag/tag.component';
-import { TagDefDirective } from '../tags/tags-list/tag-def.directive';
-import { TagsListComponent } from '../tags/tags-list/tags-list.component';
+import { TagComponent } from '../../tags/tag/tag.component';
+import { TagDefDirective } from '../../tags/tags-list/tag-def.directive';
+import { TagsListComponent } from '../../tags/tags-list/tags-list.component';
 
 @Component({
   standalone: true,
